split coral gripper subsystem
